requirements.txt
python-dotenv==0.19.2
pandas==1.3.4
questionary==1.10.0
